+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/4.4.4...HEAD)
+## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/4.4.5...HEAD)
+
+## [4.4.5](https://github.com/alexdlaird/amazon-orders/compare/4.4.4...4.4.5) - 2026-07-15
+
+### Fixed
+
+- CLI `history` crash when using `--last-30-days` or `--last-3-months`.
+- CLI `history --start-index` now validates non-numeric input.
 
 ## [4.4.4](https://github.com/alexdlaird/amazon-orders/compare/4.4.3...4.4.4) - 2026-07-07